+++ frontend/src/pages/ImageDescriptor.tsx
