+++ frontend/src/styles/Home.css
@@ -46,6 +46,7 @@
   margin: 0 0 8px;
   font-size: clamp(28px, 4vw, 44px);
   line-height: 1.1;
+  color: purple
 }
 
 .home__sub {
